refactor(babel): Use src instead of lib in tests
